requirements.txt
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
tensorboard 
#==2.0.2
tensorflow==2.3.1
tensorflow-estimator 
#==2.0.1
imutils                   
numpy                     
opencv-python-headless
pillow